requirements.txt
asgiref==3.5.2
attrs==21.4.0
awsebcli==3.20.3
bcrypt==3.2.2
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==6.2.0
XlsxWriter==3.0.3
xlwt==1.3.0
